feat: extract AccountMenu component from NavBar
Addresses issue #192 by separating wallet/account functionality into a dedicated component.

Changes:
- Created new AccountMenu component with all wallet-related UI and logic
- Moved wallet dropdown, copy operations, session deactivation, and logout from NavBar
- Used custom WalletInfo interface matching the project's wallet structure
- NavBar now focuses purely on navigation layout

Benefits:
- Better separation of concerns
- Reusable AccountMenu component
- Simplified NavBar component
- Easier to test individual components
