+++ src/components/NavBar.tsx
@@ -1,14 +1,14 @@
 'use client';
 
-import React, { useRef } from 'react';
-import { Box, Flex, Button, HStack, Text, useColorMode, Badge, Tooltip, Spinner, Image, useClipboard, Menu, MenuButton, MenuList, MenuItem, MenuDivider, useDisclosure } from '@chakra-ui/react';
-import { CopyIcon, CheckIcon, ChevronDownIcon } from '@chakra-ui/icons';
+import React from 'react';
+import { Box, Flex, Image, useColorMode } from '@chakra-ui/react';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import { useWallet } from '@/providers/WalletProvider';
 import { useSimplifiedGameState } from '@/hooks/game/useSimplifiedGameState';
 import { useSessionFunding } from '@/hooks/session/useSessionFunding';
 import { useAuthState } from '@/contexts/AuthStateContext';
+import AccountMenu from '@/components/AccountMenu';
 
 const NavBar: React.FC = () => {
   const { colorMode } = useColorMode();
@@ -33,19 +33,6 @@ const NavBar: React.FC = () => {
   const sessionKeyAddress = sessionKeyData?.key;
   const canDeactivate = !!sessionKeyAddress && sessionKeyAddress !== '0x0000000000000000000000000000000000000000';
 
-  const { onCopy: onCopyEmbedded, hasCopied: hasCopiedEmbedded } = useClipboard(embeddedWallet?.address ?? '');
-  const { onCopy: onCopyInjected, hasCopied: hasCopiedInjected } = useClipboard(injectedWallet?.address ?? '');
-
-  
-  const formatWalletType = (type?: string): string => {
-    if (!type) return '';
-    
-    return type
-      .split('_')
-      .map(word => word.charAt(0).toUpperCase() + word.slice(1))
-      .join(' ');
-  };
-  
   const handleLogout = async () => {
     try {
       await logout();
@@ -54,23 +41,9 @@ const NavBar: React.FC = () => {
       console.error("Logout failed:", error);
     }
   };
-
-  const menuTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-  const { isOpen, onClose, onOpen } = useDisclosure();
-
-  const handleCopyWithDelay = (copyFn: () => void) => {
-    // Clear any existing timeout
-    if (menuTimeoutRef.current) {
-      clearTimeout(menuTimeoutRef.current);
-    }
-    
-    // Execute the copy
-    copyFn();
-    
-    // Set a timeout to close the menu after delay
-    menuTimeoutRef.current = setTimeout(() => {
-      onClose();
-    }, 450); // 1 second delay
+  
+  const handleDeactivateSession = () => {
+    deactivateKey(undefined);
   };
 
   return (
@@ -100,122 +73,14 @@ const NavBar: React.FC = () => {
 
         <Flex justifyContent="flex-end" alignItems="center">
           {authState.hasWallet && (
-            <Menu isOpen={isOpen} onClose={onClose} onOpen={onOpen}>
-              <MenuButton as={Button} rightIcon={<ChevronDownIcon />} variant="outline" size="md">
-                {injectedWallet?.address && (
-                  <HStack spacing={1}>
-                    <Image 
-                      src="/assets/icons/avatar.png" 
-                      alt="User Avatar" 
-                      boxSize="20px" 
-                      borderRadius="full"
-                      mr={1}
-                    />
-                    <Text fontSize="sm" className='gold-text tracking-tight'>
-                      {`${injectedWallet.address.slice(0, 6)}...${injectedWallet.address.slice(-4)}`}
-                    </Text>
-                  </HStack>
-                )}
-              </MenuButton>
-              <MenuList className='!bg-transparent backdrop-blur-xl'>
-                {injectedWallet?.address && (
-                  <Tooltip 
-                    label={
-                      <HStack spacing={1}>
-                        <Text>{hasCopiedInjected ? 'Copied!' : `Copy ${formatWalletType(injectedWallet.walletClientType)}`}</Text>
-                        {hasCopiedInjected ? <CheckIcon color="green.500" /> : <CopyIcon />}
-                      </HStack>
-                    }
-                    closeOnClick={false}
-                    placement="right"
-                    hasArrow
-                  >
-                    <MenuItem 
-                      onClick={() => handleCopyWithDelay(onCopyInjected)}
-                      closeOnSelect={false}
-                      className='!bg-transparent'
-                    >
-                      <HStack spacing={1}>
-                        <Badge colorScheme="blue" fontSize="xs">{formatWalletType(injectedWallet.walletClientType).toUpperCase()}</Badge>
-                        <Text fontSize="sm" fontFamily="monospace">
-                          {`${injectedWallet.address.slice(0, 6)}...${injectedWallet.address.slice(-4)}`}
-                        </Text>
-                        {hasCopiedInjected ? (
-                          <CheckIcon color="green.500" ml={1} transition="all 0.2s ease-in-out" />
-                        ) : (
-                          <CopyIcon ml={1} transition="all 0.2s ease-in-out" />
-                        )}
-                      </HStack>
-                    </MenuItem>
-                  </Tooltip>
-                )}
-                
-                {embeddedWallet?.address && (
-                  <Tooltip 
-                    label={
-                      <HStack spacing={1}>
-                        <Text>{hasCopiedEmbedded ? 'Copied!' : 'Copy Session Key'}</Text>
-                        {hasCopiedEmbedded ? <CheckIcon color="green.500" /> : <CopyIcon />}
-                      </HStack>
-                    }
-                    closeOnClick={false}
-                    placement="right"
-                    hasArrow
-                  >
-                    <MenuItem 
-                      onClick={() => handleCopyWithDelay(onCopyEmbedded)}
-                      closeOnSelect={false}
-                      className='!bg-transparent'
-                    >
-                      <HStack spacing={1}>
-                        <Badge colorScheme="green" fontSize="xs">SESSION KEY</Badge>
-                        <Text fontSize="sm" fontFamily="monospace">
-                          {`${embeddedWallet.address.slice(0, 6)}...${embeddedWallet.address.slice(-4)}`}
-                        </Text>
-                        {hasCopiedEmbedded ? (
-                          <CheckIcon color="green.500" ml={1} transition="all 0.2s ease-in-out" />
-                        ) : (
-                          <CopyIcon ml={1} transition="all 0.2s ease-in-out" />
-                        )}
-                      </HStack>
-                    </MenuItem>
-                  </Tooltip>
-                )}
-                
-                <MenuDivider />
-                
-                <Tooltip 
-                  label="Disable session key for game actions"
-                  placement="right"
-                  hasArrow
-                >
-                  <MenuItem 
-                    onClick={() => deactivateKey(undefined)} 
-                    isDisabled={!canDeactivate || isDeactivating}
-                    closeOnSelect={false}
-                    className='!bg-transparent'
-                  >
-                    <HStack>
-                      <Text>Deactivate Session</Text>
-                      {isDeactivating && <Spinner size="sm" ml={2} />}
-                    </HStack>
-                  </MenuItem>
-                </Tooltip>
-                
-                <Tooltip 
-                  label="Disconnect your wallet"
-                  placement="right"
-                  hasArrow
-                >
-                  <MenuItem onClick={handleLogout}
-                    className='!bg-transparent'
-
-                  >
-                    Disconnect
-                  </MenuItem>
-                </Tooltip>
-              </MenuList>
-            </Menu>
+            <AccountMenu
+              injectedWallet={injectedWallet || undefined}
+              embeddedWallet={embeddedWallet || undefined}
+              canDeactivate={canDeactivate}
+              isDeactivating={isDeactivating}
+              onDeactivateSession={handleDeactivateSession}
+              onLogout={handleLogout}
+            />
           )}
         </Flex>
       </Flex>
